Reorganize authn and authz support code.
