+++ app/lib/authz-middleware.js
@@ -8,6 +8,9 @@ exports.roles = {
     visitor: 'visitor'
 }
 
+/**
+ * This middleware function verifies that a request is authorized.
+ */
 exports.requireRole = function(roles) {
     return function(req, res, next) {
         if (!req.user) {
@@ -22,4 +25,3 @@ exports.requireRole = function(roles) {
         }
     }
 }
-
