+++ app/routes/session-routes.js
@@ -2,13 +2,13 @@
 
 const express = require('express');
 const sessionController = require('../controllers/session-controller');
-const authnService = require('../services/authn-service');
+const authn = require('../lib/authn-middleware');
 
 const router = express.Router();
 
 router.route('/session')
     .get(
-        authnService.authenticate,
+        authn.authenticate,
         sessionController.retrieveCurrentSession
     );
