+++ app/routes/user-accounts-routes.js
@@ -2,46 +2,46 @@
 
 const express = require('express');
 const userAccountsController = require('../controllers/user-accounts-controller');
-const authnService = require('../services/authn-service');
-const authzService = require('../services/authz-service');
+const authn = require('../lib/authn-middleware');
+const authz = require('../lib/authz-middleware');
 
 const router = express.Router();
 
-const adminRole = [ authzService.roles.admin ];
+const adminRole = [ authz.roles.admin ];
 
 router.route('/user-accounts')
     .get(
-        authnService.authenticate,
-        authzService.requireRole(adminRole),
+        authn.authenticate,
+        authz.requireRole(adminRole),
         userAccountsController.retrieveAll
     )
     .post(
-        authnService.authenticate,
-        authzService.requireRole(adminRole),
+        authn.authenticate,
+        authz.requireRole(adminRole),
         userAccountsController.create
     );
 
 router.route('/user-accounts/:id')
     .get(
-        authnService.authenticate,
-        authzService.requireRole(adminRole),
+        authn.authenticate,
+        authz.requireRole(adminRole),
         userAccountsController.retrieveById
     )
     .put(
-        authnService.authenticate,
-        authzService.requireRole(adminRole),
+        authn.authenticate,
+        authz.requireRole(adminRole),
         userAccountsController.updateFull
     )
     .delete(
-        authnService.authenticate,
-        authzService.requireRole(adminRole),
+        authn.authenticate,
+        authz.requireRole(adminRole),
         userAccountsController.delete
     );
 
 router.route('/user-accounts/register')
     .post(
-        authnService.authenticate,
-        authzService.requireRole(adminRole),
+        authn.authenticate,
+        authz.requireRole(adminRole),
         userAccountsController.register
     );
